moved BaseProvider to prevent cyclic imports
