+++ simplemind/providers/__init__.py
@@ -1,6 +1,8 @@
+from typing import List, Type
 from .anthropic import Anthropic
 from .groq import Groq
 from .openai import OpenAI
 from .xai import XAI
+from ._base import BaseProvider
 
-providers = [Anthropic, Groq, OpenAI, XAI]
+providers: List[Type[BaseProvider]] = [Anthropic, Groq, OpenAI, XAI]
